+++ DataDefinitions/Ship.cs
@@ -91,7 +91,13 @@ namespace EddiDataDefinitions
             }
         }
 
-        /// <summary>the name of the system in which this ship is stored; null if the commander is in this ship</summary>
+        /// <summary>
+        /// The raw JSON from the companion API for this ship
+        /// </summary>
+        [JsonIgnore]
+        public string json { get; set; }
+        
+            /// <summary>the name of the system in which this ship is stored; null if the commander is in this ship</summary>
         [JsonIgnore]
         public string starsystem { get; set; }
         /// <summary>the name of the station in which this ship is stored; null if the commander is in this ship</summary>