+++ VoiceAttackResponder/VoiceAttackPlugin.cs
@@ -311,7 +311,9 @@ namespace EddiVoiceAttackResponder
             }
             else
             {
-                Process.Start(uri);
+                ProcessStartInfo proc = new ProcessStartInfo(Net.GetDefaultBrowserPath(), "\"" + uri + "\"");
+                proc.UseShellExecute = false;
+                Process.Start(proc);
             }
         }